Initialize Basic SDL Project
This started as just a means for me to get a feel for SDL so I didn't
initially plan to commit anything. However, I've decided I use what
I've created so far as a basis for a simple 2d-platformer. So
far, I load a sprite onto screen, animate it, and can run left and
right.